Calculate dmg and xp based on coverage

+++ MonsterForestDGX/Assets/PatternSpell.cs
@@ -4,7 +4,7 @@ using System;
 [Serializable]
 public class PatternSpell : ScriptableObject
 {
-    public int xp = 0;
+    public float xp = 0;
     public int level = 1;
     public int maxLevel = 4;
 }
